# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CampusConnect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: getPosts
      tags: [posts]
      summary: Get social feed posts
      security:
        - bearerAuth: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Posts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsResponse"

    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /posts/{postId}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like or unlike a post
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
  /posts/{postId}/comments:
    get:
      operationId: getComments
      tags: [posts]
      summary: Get comments for a post
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentsResponse"

    post:
      operationId: createComment
      tags: [posts]
      summary: Create a comment
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  # Users
...
  /users/{userId}/posts:
    get:
      operationId: getUserPosts
      tags: [users]
      summary: Get posts by user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # Auth schemas
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avatar:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        college:
          type: string
          nullable: true
        program:
          type: string
          nullable: true
        role:
          type: string
          enum: [student, provider, admin]
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - followersCount
        - followingCount
        - postsCount
        - createdAt
...
    Post:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        mediaUrls:
          type: array
          items:
            type: string
        author:
          $ref: "#/components/schemas/User"
        likesCount:
          type: integer
        commentsCount:
          type: integer
        isLiked:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - mediaUrls
        - author
        - likesCount
        - commentsCount
        - isLiked
        - createdAt
    PostsResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        nextCursor:
          type: string
          nullable: true
      required:
        - posts
    CreatePostRequest:
      type: object
      properties:
        content:
          type: string
        mediaUrls:
          type: array
          items:
            type: string
      required:
        - content
    LikeResponse:
      type: object
      properties:
        liked:
          type: boolean
        likesCount:
          type: integer
      required:
        - liked
        - likesCount
    Comment:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        author:
          $ref: "#/components/schemas/User"
        postId:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - author
        - postId
        - createdAt
    CommentsResponse:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
      required:
        - comments
    CreateCommentRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    # Chat schemas
...